Setup IdentityServer super basic
