+++ src/Services/Identity/Identity.api/Config.cs
@@ -1,0 +1,49 @@
+﻿using IdentityModel;
+using IdentityServer4.Models;
+using IdentityServer4.Test;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Claims;
+using System.Threading.Tasks;
+
+namespace Identity.api
+{
+    internal class Config
+    {
+        public static IEnumerable<IdentityResource> GetIdentityResources()
+        {
+            return new List<IdentityResource> {
+                new IdentityResources.OpenId(),
+                new IdentityResources.Profile(),
+                new IdentityResources.Email(),
+                new IdentityResource {
+                    Name = "role",
+                    UserClaims = new List<string> {"role"}
+                }
+            };
+        }
+
+        public static IEnumerable<ApiResource> GetApiResources()
+        {
+            return new List<ApiResource> {
+                new ApiResource {
+                    Name = "customAPI",
+                    DisplayName = "Custom API"
+                }
+            };
+        }
+
+        public static IEnumerable<Client> GetClients()
+        {
+            return new List<Client> {
+                new Client {
+                    ClientId = "client",
+                    AllowedGrantTypes = GrantTypes.ClientCredentials,
+                    ClientSecrets = { new Secret("secret".Sha256()) },
+                    AllowedScopes = { "api1" }
+                }
+            };
+        }
+    }
+}
