--- src/Services/Identity/Identity.api/IdentityServerConfiguration.cs
@@ -1,72 +1,0 @@
-﻿using IdentityModel;
-using IdentityServer4.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Claims;
-using System.Threading.Tasks;
-
-namespace Identity.api
-{
-    internal class IdentityServerConfiguration
-    {
-        public static IEnumerable<IdentityResource> GetIdentityResources()
-        {
-            return new List<IdentityResource> {
-                new IdentityResources.OpenId(),
-                new IdentityResources.Profile(),
-                new IdentityResources.Email(),
-                new IdentityResource {
-                    Name = "role",
-                    UserClaims = new List<string> {"role"}
-                }
-            };
-        }
-
-        public static IEnumerable<ApiResource> GetApiResources()
-        {
-            return new List<ApiResource> {
-                new ApiResource {
-                    Name = "customAPI",
-                    DisplayName = "Custom API",
-                    Description = "Custom API Access",
-                    UserClaims = new List<string> {"role"},
-                    ApiSecrets = new List<Secret> {new Secret("scopeSecret".Sha256())},
-                    Scopes = new List<Scope> {
-                        new Scope("customAPI.read"),
-                        new Scope("customAPI.write")
-                    }
-                }
-            };
-        }
-
-        public static IEnumerable<Client> GetClients()
-        {
-            return new List<Client> {
-                new Client {
-                    ClientId = "oauthClient",
-                    ClientName = "Example Client Credentials Client Application",
-                    AllowedGrantTypes = GrantTypes.ClientCredentials,
-                    ClientSecrets = new List<Secret> {
-                        new Secret("superSecretPassword".Sha256())},
-                    AllowedScopes = new List<string> {"customAPI.read"}
-                }
-            };
-        }
-
-        public static List<IdentityResource> GetIdentities()
-        {
-            var customProfile = new IdentityResource(
-                name: "custom.profile",
-                displayName: "Custom profile",
-                claimTypes: new[] { "name", "email", "status" });
-
-            return new List<IdentityResource>
-            {
-                new IdentityResources.OpenId(),
-                new IdentityResources.Profile(),
-                customProfile
-            };
-        }
-    }
-}
