+++ src/Services/Identity/Identity.api/Program.cs
@@ -14,6 +14,7 @@ namespace Identity.api
         {
             var host = new WebHostBuilder()
                 .UseKestrel()
+                .UseUrls("http://localhost:5000")
                 .UseContentRoot(Directory.GetCurrentDirectory())
                 .UseIISIntegration()
                 .UseStartup<Startup>()
